ASM-5063 FCoE Deployment with Cisco Nexus failing intermittently due to missing discovery facts
Extra "\n" was added to the SSH Send block leading to an extra prompt in the command output. As a result if any command is sending blank output then the next command output was getting appended with extra prompts.

Also added command "show version" after commands sending blank output so that further command outputs are not affected